# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/commission-works": {
      "post": {
        "description": "commissionId가 있으면 FORM(폼) 신청, 없으면 DIRECT(직접 의뢰). 신청 시 CommissionWork와 CommissionRequest가 동시에 생성됨.",
        "operationId": "commissionWork_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommissionWorkRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommissionWork400Response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommissionWork404Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "커미션 신청",
        "tags": [
          "commission-works"
        ]
      }
    },
    "/commission-works/{id}/accept": {
      "patch": {
        "description": "작가가 받은 PENDING 상태의 신청을 수락. ACCEPTED 상태로 전환.",
        "operationId": "commissionWork_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptCommissionWork403Response"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptCommissionWork404Response"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptCommissionWork409Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "커미션 수락",
        "tags": [
          "commission-works"
        ]
      }
    },
    "/commission-works/{id}/cancel": {
      "patch": {
        "description": "신청자(의뢰인)가 본인이 보낸 PENDING 상태의 신청을 취소. CANCELED 상태로 전환. 작가가 이미 수락/거절했으면 409.",
        "operationId": "commissionWork_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelCommissionWork403Response"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelCommissionWork404Response"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelCommissionWork409Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "커미션 신청 취소",
        "tags": [
          "commission-works"
        ]
      }
    },
    "/commission-works/{id}/complete": {
      "patch": {
        "description": "신청자(의뢰인)가 작업을 완료 처리. ACCEPTED/IN_PROGRESS/FINAL 상태에서 COMPLETED로 전환.",
        "operationId": "commissionWork_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteCommissionWork403Response"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteCommissionWork404Response"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteCommissionWork409Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "커미션 작업 완료",
        "tags": [
          "commission-works"
        ]
      }
    },
    "/commission-works/{id}/reject": {
      "patch": {
        "description": "작가가 받은 PENDING 상태의 신청을 거절. 거절 사유는 선택.",
        "operationId": "commissionWork_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectCommissionWorkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectCommissionWork403Response"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectCommissionWork404Response"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectCommissionWork409Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "커미션 거절",
        "tags": [
          "commission-works"
        ]
      }
    },
    "/commission-works/{id}/result": {
      "put": {
        "description": "작가가 작업물 이미지를 업로드. 기존 작업물 전체를 덮어씀. 작업물이 있으면 IN_PROGRESS, 최종 체크 시 FINAL로 전환. 종료 상태(REJECTED/CANCELED/COMPLETED)면 409.",
        "operationId": "commissionWork_uploadResult",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCommissionWorkResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCommissionWorkResult403Response"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCommissionWorkResult404Response"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCommissionWorkResult409Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "작업물 업로드 (덮어쓰기)",
        "tags": [
          "commission-works"
        ]
      }
    },
    "/commission-works/{id}/memos": {
      "post": {
        "description": "작가가 해당 커미션 작업에 메모를 작성. 메모는 의뢰인에게 노출됨. 여러 개 작성 가능.",
        "operationId": "commissionWork_createMemo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommissionWorkMemoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommissionWorkMemo403Response"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommissionWorkMemo404Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "작업 메모 작성",
        "tags": [
          "commission-works"
        ]
      }
    },
    "/commission-works/{id}/reviews": {
      "post": {
        "description": "COMPLETED 상태의 커미션에 대해 의뢰인/작가가 상대방에게 후기를 작성. 한 사람당 1회만 작성 가능.",
        "operationId": "commissionWork_createReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommissionReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommissionReview403Response"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommissionReview404Response"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommissionReview409Response"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "커미션 후기(유저 평가) 작성",
        "tags": [
          "commission-works"
        ]
      }
    }
  },
  "info": {
    "title": "grimity API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "IdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "CreateCommissionWorkRequest": {
        "type": "object",
        "properties": {
          "authorId": {
            "type": "string",
            "description": "수신자(작가) userId",
            "format": "uuid"
          },
          "referenceImages": {
            "example": [
              "v2/commission-work/{UUID}.jpg"
            ],
            "description": "레퍼런스 이미지 (0~10개)",
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 10
            }
          },
          "commissionId": {
            "type": "object",
            "nullable": true,
            "format": "uuid",
            "description": "있으면 FORM 모드(폼 신청), 없으면 DIRECT 모드(직접 의뢰)"
          },
          "answers": {
            "description": "FORM 모드: 배열 인덱스 = 질문 order, text/selectedOptions만 사용(메타는 서버가 무시). DIRECT 모드: type/title 포함한 완전한 답변 아이템.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommissionAnswerItem",
              "maxLength": 20
            }
          }
        },
        "required": [
          "authorId",
          "referenceImages"
        ]
      },
      "CreateCommissionWork400Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              400
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "SELF_REQUEST_NOT_ALLOWED",
              "COMMISSION_AUTHOR_MISMATCH",
              "ANSWERS_LENGTH_MISMATCH",
              "ANSWER_TYPE_REQUIRED",
              "ANSWER_TITLE_REQUIRED",
              "SELECT_OPTIONS_REQUIRED",
              "TEXT_ANSWER_REQUIRED",
              "TEXT_HAS_SELECTED_OPTIONS",
              "SELECT_HAS_TEXT",
              "SELECT_HAS_ATTACHED_IMAGES",
              "SELECTED_OPTION_NOT_IN_OPTIONS",
              "SELECTED_OPTIONS_DUPLICATED",
              "SINGLE_SELECT_ANSWER_INVALID",
              "MULTI_SELECT_ANSWER_REQUIRED"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CreateCommissionWork404Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              404
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "AUTHOR_NOT_FOUND",
              "COMMISSION_NOT_FOUND"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "AcceptCommissionWork403Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              403
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "NOT_COMMISSION_AUTHOR"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "AcceptCommissionWork404Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              404
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_FOUND"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "AcceptCommissionWork409Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              409
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_PENDING"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CancelCommissionWork403Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              403
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "NOT_COMMISSION_CLIENT"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CancelCommissionWork404Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              404
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_FOUND"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CancelCommissionWork409Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              409
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_PENDING"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CompleteCommissionWork403Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              403
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "NOT_COMMISSION_CLIENT"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CompleteCommissionWork404Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              404
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_FOUND"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CompleteCommissionWork409Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              409
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_COMPLETABLE"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "RejectCommissionWorkRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "object",
            "nullable": true,
            "maxLength": 500,
            "description": "거절 사유 (선택)"
          }
        }
      },
      "RejectCommissionWork403Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              403
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "NOT_COMMISSION_AUTHOR"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "RejectCommissionWork404Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              404
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_FOUND"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "RejectCommissionWork409Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              409
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_PENDING"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "UploadCommissionWorkResultRequest": {
        "type": "object",
        "properties": {
          "images": {
            "example": [
              "v2/commission-work/{UUID}.jpg"
            ],
            "description": "작업물 이미지 (1~20개)",
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 20,
              "minLength": 1
            }
          },
          "isFinal": {
            "type": "boolean",
            "description": "최종 작업물 여부"
          }
        },
        "required": [
          "images",
          "isFinal"
        ]
      },
      "UploadCommissionWorkResult403Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              403
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "NOT_COMMISSION_AUTHOR"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "UploadCommissionWorkResult404Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              404
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_FOUND"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "UploadCommissionWorkResult409Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              409
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "WORK_NOT_ACTIVE"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
      "CreateCommissionWorkMemoRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "description": "작업 메모 내용"
          }
        },
        "required": [
          "content"
        ]
      },
      "CreateCommissionWorkMemo403Response": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "enum": [
              403
            ]
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "NOT_COMMISSION_AUTHOR"
            ]
          }
        },
        "required": [
          "status",
          "errorCode"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "유효성 검사 실패 시 문자열 배열로 내려온다"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}